requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
colorlog==5.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
filetype==1.0.7
har2case==0.3.1
HttpRunner==1.5.8
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpath==0.82
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==20.9
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
ruamel.yaml.clib==0.2.2
sentry-sdk==0.13.5
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
mysqlclient==1.4.4